+++ CHANGELOG.md
@@ -6,6 +6,7 @@ for Rust libraries in [RFC #1105](https://github.com/rust-lang/rfcs/blob/master/
 
 ## [0.2.0] - (pending)
 
+ * Instead of panicing, all methods that could return a `Result<T, grok::Error>`.
  * `Grok::new()` has been renamed to `Grok::empty()` (or `Grok::default()`)
  * `is_empty()` is available in the `Matches` API to check if there are matches at all.
  * `len()` is available in the `Matches` API to get the total number of matches.